Mon, May 23, 2022, 12:13 AM +00:00

+++ filter-less/helpers.c
@@ -16,7 +16,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         {
                 // average the color values of red, green, and blue
                 // assiasgn the new average back to red, green and blue
-            float gray = 3.0 / (image[i][j].rgbtRed + image[i][j].rgbtBlue + image[i][j].rgbtBlue);
+            int gray = 3.0 / (image[i][j].rgbtRed + image[i][j].rgbtBlue + image[i][j].rgbtBlue);
             round(gray);
             // assiasgn the new average back to red, green and blue
             image[i][j].rgbtRed = gray;
